requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-config-url==0.1.1
Django==2.2
django-cors-middleware==1.5.0
django-oauth-toolkit==1.3.2
django-rest-swagger==2.2.0
django-swagger-ui==0.1.11
djangorestframework==3.11.1
future==0.18.2
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
oauthlib==3.1.0
openapi-codec==1.3.2
protobuf==3.12.4
psycopg2==2.8.5
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
swagger-ui-bundle==0.0.8
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
